requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
supabase>=2.0.0
openai>=1.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0
httpx>=0.25.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 